+++ priv/templates/html/authorize_test.exs
@@ -2,75 +2,71 @@ defmodule <%= base %>.AuthorizeTest do
   use <%= base %>.ConnCase
 
   import <%= base %>.OpenmaizeEcto
-  import OpenmaizeJWT.Create
   alias <%= base %>.{Repo, User}
 
   @valid_attrs %{email: "tony@mail.com", password: "mangoes&g0oseberries"}
   @invalid_attrs %{email: "tony@mail.com", password: "maaaangoes&g00zeberries"}
 
-  {:ok, user_token} = %{id: 3, email: "tony@mail.com", role: "user"}
-                      |> generate_token({0, 1440})
-  @user_token user_token
+  # In this example setup, `conn` is the connection for an unauthenticated
+  # user, and `user_conn` is for an authenticated user (with an id of 3)
+  # You will need a similar setup in some of your other controller files,
+  # if your tests involve the use of sessions
+  setup %{conn: conn} do
+    conn = conn |> bypass_through(<%= base %>.Router, :browser) |> get("/")
+    user_conn = conn |> put_session(:user_id, 3) |> send_resp(:ok, "/")
 
-  setup do
-    conn = build_conn()
-    |> put_req_cookie("access_token", @user_token)
-    {:ok, conn: conn}
+    {:ok, %{conn: conn, user_conn: user_conn}}
   end
 
   # The first three tests can be used to test routes protected by
   # the role_check plug or the custom action (authorize_action) function
-  test "correct user role is successfully authorized", %{conn: conn} do
-    conn = get conn, "/users"
+  test "correct user role is successfully authorized", %{user_conn: user_conn} do
+    conn = get user_conn, "/users"
     assert html_response(conn, 200)
   end
 
-  test "authorization for incorrect role fails", %{conn: conn} do
-    conn = get conn, "/admin"
+  test "authorization for incorrect role fails", %{user_conn: user_conn} do
+    conn = get user_conn, "/admin"
     assert redirected_to(conn) == "/users"
   end
 
-  test "authorization for nil user fails" do
-    conn = build_conn() |> get("/users")
+  test "authorization for nil user fails", %{conn: conn} do
+    conn = conn |> get("/users")
     assert redirected_to(conn) == "/login"
   end
 
   # Test routes protected by the id_check plug
-  test "id check succeeds", %{conn: conn} do
-    conn = get conn, "/users/3"
+  test "id check succeeds", %{user_conn: user_conn} do
+    conn = get user_conn, "/users/3"
     assert html_response(conn, 200)
   end
 
-  test "id check fails for incorrect id", %{conn: conn} do
-    conn = get conn, "/users/30"
+  test "id check fails for incorrect id", %{user_conn: user_conn} do
+    conn = get user_conn, "/users/30"
     assert redirected_to(conn) == "/users"
   end
 
-  test "id check fails for nil user" do
-    conn = build_conn() |> get("/users/3")
+  test "id check fails for nil user", %{conn: conn} do
+    conn = conn |> get("/users/3")
     assert redirected_to(conn) == "/login"
   end
 
-  test "login succeeds" do
+  test "login succeeds", %{conn: conn} do
     # Remove the Repo.get_by line if you are not using email confirmation
     Repo.get_by(User, %{email: "tony@mail.com"}) |> user_confirmed
-    conn = post build_conn(), "/login", user: @valid_attrs
+    conn = post conn, "/login", user: @valid_attrs
     assert redirected_to(conn) == "/users"
   end
 
-  test "login fails" do
+  test "login fails", %{conn: conn} do
     # Remove the Repo.get_by line if you are not using email confirmation
     Repo.get_by(User, %{email: "reg@mail.com"}) |> user_confirmed
-    conn = post build_conn(), "/login", user: @invalid_attrs
+    conn = post conn, "/login", user: @invalid_attrs
     assert redirected_to(conn) == "/login"
   end
 
-  test "logout succeeds" do
-  {:ok, user_token} = %{id: 3, email: "tony@mail.com", role: "user"}
-                      |> generate_token({0, 1440})
-    conn = build_conn()
-    |> put_req_cookie("access_token", user_token)
-    |> delete("/logout")
+  test "logout succeeds", %{user_conn: user_conn} do
+    conn = delete user_conn, "/logout"
     assert redirected_to(conn) == "/"
   end